+++ src/Axent.Abstractions/IPipelineExecutorService.cs
@@ -1,8 +1,0 @@
-﻿namespace Axent.Abstractions;
-
-public interface IPipelineExecutorService
-{
-    ValueTask<Response<TResponse>> ExecuteAsync<TRequest, TResponse>(IEnumerable<IAxentPipe<TRequest, TResponse>> pipes,
-        RequestContext<TRequest> request,
-        CancellationToken cancellationToken = default);
-}