+++ src/Axent.Core/AxentOptions.cs
@@ -3,8 +3,18 @@ namespace Axent.Core;
 public sealed class AxentOptions
 {
     /// <summary>
-    /// Determines whether to use the source-generated sender implementation.
-    /// Defaults to true.
+    /// Determines whether error handling is enabled or exceptions should be "forwarded" to the consumer
+    /// of the library.
+    /// No errors will be caught if the options are not set.
     /// </summary>
-    public bool UseSourceGeneratedSender { get; set; } = true;
+    public AxentErrorHandlingOptions? ErrorHandling { get; set; }
+}
+
+public sealed class AxentErrorHandlingOptions
+{
+    /// <summary>
+    /// Determines whether the full exception gets returned or not, when the `ErrorHandlingPipe` is registered.
+    /// Defaults to false
+    /// </summary>
+    public bool EnableDetailedExceptionResponse { get; set; }
 }